v2: fix json encoding bug for job get JOB_UUID -o json
The tabler.HTCJob interface hides the concrete type's UnmarshalJSON
method, making it so that json unmarshaling goes whole hog on encoding
properties like instanceId that are not necessarily set (e.g. in Azure
right now), and thus making the JSON encoder fail.

* Fix this temporarily by giving HTCJob a MarshalJSON method that calls
  down into its concrete type.
* Add TODO with a way to resolve this more robustly going forward.

In addition, this also fixes json and yaml output so that when an empty
or nil slice has been passed to PrintResult (even if it's wrapped up in
a Tabler interface), we don't print `[]` or worse yet, `null`, but
instead just print nothing.